+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2.0.3
+
+Bugfixes:
+* Fix typo in config parsing for --color-title (#534)
+* Fix percent formatting for `--*-format` (#535)
+* Fix loading presets for homebrew (macOS)
+
+Features:
+* Add option `--percent-ndigits`
+* Add command flag `--config` as an alias of `--load-config`
+* Windows packages now include presets (Windows)
+
 # 2.0.2
 
 Bugfixes: